requirements.txt
Flask==3.1.0
gunicorn==23.0.0
google-generativeai